requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
safetensors>=0.3.0
pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.0
moviepy>=1.0.3
tqdm>=4.65.0
pyyaml>=6.0
huggingface-hub>=0.16.0
einops>=0.6.1
xformers>=0.0.20
peft>=0.4.0

